% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/d0.R
\name{d0}
\alias{d0}
\title{d0 Function to run limma voom edgeR on se object.}
\usage{
d0(se)
}
\arguments{
\item{se}{SummExp object.}
}
\value{
Returns/prints voom plots.
}
\description{
This function runs limma voom edgeR on se object.
}
\details{
This function runs limma voom edgeR on se object.
}
\examples{
type1 <- simData()
d0(type1)
}
\references{
The 'doseLM' package, 2019 (in press).
}
\author{
AJ Vaestermark, JR Walters.
}
