% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/d1.R
\name{d1}
\alias{d1}
\title{d1 Function to run Tom's ANOVA on se object.}
\usage{
d1(se)
}
\arguments{
\item{se}{SummExp object.}
}
\value{
Returns/prints Tom's ANOVA.
}
\description{
This function runs Tom's ANOVA on se object.
}
\details{
This function runs Tom's ANOVA on se object.
}
\examples{
type1 <- simData()
d1(type1)
}
\references{
The 'doseLM' package, 2019 (in press).
}
\author{
AJ Vaestermark, JR Walters.
}
